+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [v0.10.8](https://github.com/NubeIO/flow-framework/tree/v0.10.8) (2023-03-08)
+
+- Implementation of MQTT publish buffer
+
+## [v0.10.7](https://github.com/NubeIO/flow-framework/tree/v0.10.7) (2023-03-08)
+
+- Fix: point write issue from MQTT
+
 ## [v0.10.6](https://github.com/NubeIO/flow-framework/tree/v0.10.6) (2023-03-07)
 
 - Fix: points values are not written correctly